Migrate QKSMS+ screen to controller
