+++ presentation/src/main/java/com/moez/QKSMS/feature/plus/PlusActivity.kt
@@ -18,100 +18,40 @@
  */
 package com.moez.QKSMS.feature.plus
 
-import android.graphics.Typeface
 import android.os.Bundle
-import androidx.core.view.children
-import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.ViewModelProviders
-import com.jakewharton.rxbinding2.view.clicks
-import com.moez.QKSMS.BuildConfig
+import androidx.core.view.isVisible
+import com.bluelinelabs.conductor.Conductor
+import com.bluelinelabs.conductor.Router
+import com.bluelinelabs.conductor.RouterTransaction
 import com.moez.QKSMS.R
 import com.moez.QKSMS.common.base.QkThemedActivity
-import com.moez.QKSMS.common.util.BillingManager
-import com.moez.QKSMS.common.util.FontProvider
-import com.moez.QKSMS.common.util.extensions.resolveThemeColor
-import com.moez.QKSMS.common.util.extensions.setBackgroundTint
-import com.moez.QKSMS.common.util.extensions.setTint
-import com.moez.QKSMS.common.util.extensions.setVisible
 import com.moez.QKSMS.common.util.extensions.viewBinding
-import com.moez.QKSMS.common.widget.PreferenceView
-import com.moez.QKSMS.databinding.QksmsPlusActivityBinding
-import com.moez.QKSMS.feature.plus.experiment.UpgradeButtonExperiment
+import com.moez.QKSMS.databinding.ContainerActivityBinding
+import com.moez.QKSMS.feature.contacts.ContactsActivity
+import com.moez.QKSMS.feature.contacts.ContactsController
 import dagger.android.AndroidInjection
-import javax.inject.Inject
 
-class PlusActivity : QkThemedActivity(), PlusView {
+class PlusActivity : QkThemedActivity() {
 
-    @Inject lateinit var fontProvider: FontProvider
-    @Inject lateinit var upgradeButtonExperiment: UpgradeButtonExperiment
-    @Inject lateinit var viewModelFactory: ViewModelProvider.Factory
-
-    private val binding by viewBinding(QksmsPlusActivityBinding::inflate)
-    private val viewModel by lazy { ViewModelProviders.of(this, viewModelFactory)[PlusViewModel::class.java] }
-
-    override val upgradeIntent by lazy { binding.upgrade.clicks() }
-    override val upgradeDonateIntent by lazy { binding.upgradeDonate.clicks() }
-    override val donateIntent by lazy { binding.donate.clicks() }
-    override val themeClicks by lazy { binding.themes.clicks() }
-    override val scheduleClicks by lazy { binding.schedule.clicks() }
-    override val backupClicks by lazy { binding.backup.clicks() }
-    override val delayedClicks by lazy { binding.delayed.clicks() }
-    override val nightClicks by lazy { binding.night.clicks() }
+    private val binding by viewBinding(ContainerActivityBinding::inflate)
+    private lateinit var router: Router
 
     override fun onCreate(savedInstanceState: Bundle?) {
         AndroidInjection.inject(this)
         super.onCreate(savedInstanceState)
         setContentView(binding.root)
-        setTitle(R.string.title_qksms_plus)
-        showBackButton(true)
-        viewModel.bindView(this)
+        toolbar.isVisible = false
 
-        binding.free.setVisible(false)
-
-        if (!prefs.systemFont.get()) {
-            fontProvider.getLato { lato ->
-                val typeface = Typeface.create(lato, Typeface.BOLD)
-                binding.appBarLayout.collapsingToolbar.setCollapsedTitleTypeface(typeface)
-                binding.appBarLayout.collapsingToolbar.setExpandedTitleTypeface(typeface)
-            }
+        router = Conductor.attachRouter(this, binding.container, savedInstanceState)
+        if (!router.hasRootController()) {
+            router.setRoot(RouterTransaction.with(PlusController()))
         }
-
-        // Make the list titles bold
-        binding.linearLayout.children
-                .mapNotNull { view -> view as? PreferenceView }
-                .map { preferenceView -> preferenceView.binding.titleView }
-                .forEach { it.setTypeface(it.typeface, Typeface.BOLD) }
-
-        val textPrimary = resolveThemeColor(android.R.attr.textColorPrimary)
-        binding.appBarLayout.collapsingToolbar.setCollapsedTitleTextColor(textPrimary)
-        binding.appBarLayout.collapsingToolbar.setExpandedTitleColor(textPrimary)
-
-        val theme = colors.theme().theme
-        binding.donate.setBackgroundTint(theme)
-        binding.upgrade.setBackgroundTint(theme)
-        binding.thanksIcon.setTint(theme)
     }
 
-    override fun render(state: PlusState) {
-        binding.description.text = getString(R.string.qksms_plus_description_summary, state.upgradePrice)
-        binding.upgrade.text = getString(upgradeButtonExperiment.variant, state.upgradePrice, state.currency)
-        binding.upgradeDonate.text = getString(R.string.qksms_plus_upgrade_donate, state.upgradeDonatePrice, state.currency)
-
-        val fdroid = BuildConfig.FLAVOR == "noAnalytics"
-
-        binding.free.setVisible(fdroid)
-        binding.toUpgrade.setVisible(!fdroid && !state.upgraded)
-        binding.upgraded.setVisible(!fdroid && state.upgraded)
-
-        binding.themes.isEnabled = state.upgraded
-        binding.schedule.isEnabled = state.upgraded
-        binding.backup.isEnabled = state.upgraded
-        binding.delayed.isEnabled = state.upgraded
-        binding.night.isEnabled = state.upgraded
-    }
-
-    override fun initiatePurchaseFlow(billingManager: BillingManager, sku: String) {
-        billingManager.initiatePurchaseFlow(this, sku)
+    override fun onBackPressed() {
+        if (!router.handleBack()) {
+            super.onBackPressed()
+        }
     }
 
 }
