--- presentation/src/main/java/com/moez/QKSMS/feature/plus/PlusActivityModule.kt
@@ -1,35 +1,0 @@
-/*
- * Copyright (C) 2017 Moez Bhatti <moez.bhatti@gmail.com>
- *
- * This file is part of QKSMS.
- *
- * QKSMS is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * QKSMS is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with QKSMS.  If not, see <http://www.gnu.org/licenses/>.
- */
-package com.moez.QKSMS.feature.plus
-
-import androidx.lifecycle.ViewModel
-import com.moez.QKSMS.injection.ViewModelKey
-import dagger.Module
-import dagger.Provides
-import dagger.multibindings.IntoMap
-
-@Module
-class PlusActivityModule {
-
-    @Provides
-    @IntoMap
-    @ViewModelKey(PlusViewModel::class)
-    fun providePlusViewModel(viewModel: PlusViewModel): ViewModel = viewModel
-
-}
