+++ presentation/src/main/java/com/moez/QKSMS/feature/plus/PlusView.kt
@@ -18,11 +18,11 @@
  */
 package com.moez.QKSMS.feature.plus
 
-import com.moez.QKSMS.common.base.QkView
+import com.moez.QKSMS.common.base.QkViewContract
 import com.moez.QKSMS.common.util.BillingManager
 import io.reactivex.Observable
 
-interface PlusView : QkView<PlusState> {
+interface PlusView : QkViewContract<PlusState> {
 
     val upgradeIntent: Observable<Unit>
     val upgradeDonateIntent: Observable<Unit>
